reorganize validation helpers
