+++ src/common/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod bool_or;
 pub mod extensions;
+pub mod helpers;
 pub mod reference;
