+++ src/common/reference.rs
@@ -2,9 +2,10 @@
 
 use std::ops::Add;
 
+use crate::common::helpers::{Context, ValidateWithContext};
 use serde::{Deserialize, Serialize};
 
-use crate::validation::{Context, Options, ValidateWithContext};
+use crate::validation::Options;
 
 pub enum ResolveError {
     NotFound,
