+++ src/v2/parameter.rs
@@ -3,6 +3,7 @@
 use std::collections::BTreeMap;
 use std::ops::Add;
 
+use crate::common::helpers::{validate_required_string, Context, ValidateWithContext};
 use serde::{Deserialize, Serialize};
 
 use crate::common::reference::RefOr;
@@ -10,7 +11,6 @@ use crate::v2::formats::{CollectionFormat, IntegerFormat, NumberFormat, StringFo
 use crate::v2::items::Items;
 use crate::v2::schema::Schema;
 use crate::v2::spec::Spec;
-use crate::validation::{validate_required_string, Context, ValidateWithContext};
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
 #[serde(tag = "in")]
