+++ src/v2/xml.rs
@@ -1,10 +1,12 @@
 use std::collections::BTreeMap;
 use std::ops::Add;
 
+use crate::common::helpers::{
+    validate_required_string, validate_url, Context, ValidateWithContext,
+};
 use serde::{Deserialize, Serialize};
 
 use crate::v2::spec::Spec;
-use crate::validation::{validate_required_string, validate_url, Context, ValidateWithContext};
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Default)]
 pub struct XML {
